--- src/js/artists.js
@@ -1,1 +1,0 @@
-
