--- src/js/artists1.js
@@ -1,2 +1,0 @@
-
-import "./artists/index.js";
